requirements.txt
streamlit==1.37.0
pypdf==4.2.0
faiss-cpu==1.13.2
sentence-transformers==3.0.1
google-generativeai==0.7.2
reportlab==4.2.2
numpy==1.26.4
scikit-learn==1.5.1
tqdm==4.66.4
python-dotenv==1.0.0
